requirements.txt
python-dotenv==1.2.1
playwright==1.55.0
pytest-playwright==0.7.2
